Added hostname as devicename + added cancel event in browser
