+++ Auth0.Windows/Auth0.Windows.csproj
@@ -87,7 +87,8 @@
     <Compile Include="BrowserAuthenticationForm.Designer.cs">
       <DependentUpon>BrowserAuthenticationForm.cs</DependentUpon>
     </Compile>
-    <Compile Include="Device.cs" />
+    <Compile Include="DeviceIdProvider.cs" />
+    <Compile Include="Extensions.cs" />
     <Compile Include="IDeviceIdProvider.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="TokenValidator.cs" />
