--- Auth0.Windows/Device.cs
@@ -1,10 +1,0 @@
-﻿namespace Auth0.Windows
-{
-    public class Device : IDeviceIdProvider
-    {
-        public string GetDeviceId()
-        {
-            return "Windows Device";
-        }
-    }
-}
