+++ Auth0.Windows/DeviceIdProvider.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace Auth0.Windows
+{
+    public class DeviceIdProvider : IDeviceIdProvider
+    {
+        public string GetDeviceId()
+        {
+            return Environment.MachineName;
+        }
+    }
+}
