+++ Auth0.Windows/Extensions.cs
@@ -1,0 +1,12 @@
+using Newtonsoft.Json.Linq;
+
+namespace Auth0.Windows
+{
+    internal static class Extensions
+    {
+        internal static JObject ToJson(this string jsonString)
+        {
+            return JObject.Parse(jsonString);
+        }
+    }
+}
