+++ Auth0.Windows/IDeviceIdProvider.cs
@@ -1,6 +1,4 @@
-﻿using System.Threading.Tasks;
-
-namespace Auth0.Windows
+﻿namespace Auth0.Windows
 {
     public interface IDeviceIdProvider
     {
